# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi Documentation",
        "description": "Documentation for verde inc",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Base API"
        }
    ],
    "paths": {
        "/api/auth/agents/register": {
            "post": {
                "tags": [
                    "Agents"
                ],
                "summary": "Register for agents",
                "operationId": "registerAgents",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "description login agents",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "testLogin"
                    },
                    {
                        "name": "surname",
                        "in": "path",
                        "description": "description surname agents",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "testSurname"
                    },
                    {
                        "name": "birthdate",
                        "in": "path",
                        "description": "description age agents",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "18/07/1998"
                    },
                    {
                        "name": "email",
                        "in": "path",
                        "description": "description email agents",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "testEmail@gmail.com"
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "description password agents",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "testPassword"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "You have successfully registered"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "testName"
                                                    },
                                                    "surname": {
                                                        "type": "string",
                                                        "example": "surnameTest"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "test@mail.ru"
                                                    },
                                                    "email_verified_at": {
                                                        "type": "string",
                                                        "example": "null | 2023-01-22 23:10:27"
                                                    },
                                                    "date_of_birth": {
                                                        "type": "string",
                                                        "example": "null | 1998-18-12"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-01-22 23:10:27"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-01-22 23:10:27"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/agents/login": {
            "post": {
                "tags": [
                    "Agents"
                ],
                "summary": "Login for agents",
                "operationId": "loginAgents",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "description email agents",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "testEmail@gmail.com"
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "description password agents",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "testPassword"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "success": {
                                                "type": "boolean"
                                            },
                                            "bearerToken": {
                                                "type": "string",
                                                "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9sLnZlcmRlLmxvY1wvYXBpXC9hdXRoXC9hZ2VudHNcL2xvZ2luIiwiaWF0IjoxNjc0NDMxODg4LCJleHAiOjE2NzQ0MzU0ODgsIm5iZiI6MTY3NDQzMTg4OCwianRpIjoiTUN1N2dzcXllREdVUVNaMSIsInN1YiI6MTUsInBydiI6IjczM2ZkNzg0Yzc4NmI3ZTUwM2FjZmU0ZGFhOTMwMDNjZWRkNmVlYmIifQ.jSOnS-ydYXc53KS0URwxNHY0Gy43YU8EEf9oSSvlSKY"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "You have successfully logged in to your account."
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/agents/profile": {
            "post": {
                "tags": [
                    "Agents"
                ],
                "summary": "Getting profile data for an agent",
                "description": "Protected bearer-token",
                "operationId": "profileCatalog",
                "parameters": [
                    {
                        "name": "bearer",
                        "in": "path",
                        "description": "bearer-token user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9sLnZlcmRlLmxvY1wvYXBpXC9hdXRoXC9hZ2VudHNcL2xvZ2luIiwiaWF0IjoxNjc0NDMxODg4LCJleHAiOjE2NzQ0MzU0ODgsIm5iZiI6MTY3NDQzMTg4OCwianRpIjoiTUN1N2dzcXllREdVUVNaMSIsInN1YiI6MTUsInBydiI6IjczM2ZkNzg0Yzc4NmI3ZTUwM2FjZmU0ZGFhOTMwMDNjZWRkNmVlYmIifQ.jSOnS-ydYXc53KS0URwxNHY0Gy43YU8EEf9oSSvlSKY"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Your personal data"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "testName"
                                                    },
                                                    "surname": {
                                                        "type": "string",
                                                        "example": "surnameTest"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "test@mail.ru"
                                                    },
                                                    "email_verified_at": {
                                                        "type": "string",
                                                        "example": "null | 2023-01-22 23:10:27"
                                                    },
                                                    "date_of_birth": {
                                                        "type": "string",
                                                        "example": "null | 1998-18-12"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-01-22 23:10:27"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-01-22 23:10:27"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/agents/logout": {
            "post": {
                "tags": [
                    "Agents"
                ],
                "summary": "logoutAgents for agents",
                "operationId": "logoutAgents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "success": {
                                                "type": "boolean"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "You have successfully logged out of your account"
                                            },
                                            "data": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "logout": {
                                                            "type": "string",
                                                            "example": "success"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}